Move GH search request to server
Instead of rendering empty and then searching in an effect, we can move the search request to the server to SSR the complete site.
